requirements.txt
grpcio==1.75.1
grpcio-tools==1.75.1
numpy==2.3.3
protobuf==6.32.1
setuptools==80.9.0
typing_extensions==4.15.0
